Refactor code and add more tests
